+++ android/src/main/kotlin/com/apphud/fluttersdk/ApphudPlugin.kt
@@ -4,7 +4,7 @@ import android.app.Activity
 import android.content.Context
 import androidx.annotation.NonNull
 import com.apphud.fluttersdk.handlers.*
-import com.apphud.sdk.client.HttpUrlConnectionExecutor
+import com.apphud.sdk.managers.HeadersInterceptor
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodCall
@@ -27,11 +27,15 @@ class ApphudPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
     private var activity: Activity? = null
 
     private lateinit var channel: MethodChannel
+    private lateinit var listenerChannel: MethodChannel
+    private lateinit var listenerHandler: ApphudListenerHandler
 
     override
     fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
         channel = MethodChannel(flutterPluginBinding.binaryMessenger, "apphud")
         channel.setMethodCallHandler(this)
+        listenerChannel = MethodChannel(flutterPluginBinding.binaryMessenger, "apphud/listener")
+        listenerHandler = ApphudListenerHandler(listenerChannel)
         this.context = flutterPluginBinding.applicationContext
     }
 
@@ -39,14 +43,17 @@ class ApphudPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
     override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
         handlers.forEach { handler ->
             if (handler.isAbleToHandle(call.method)) handler.tryToHandle(
-                    method = call.method,
-                    args = call.arguments as? Map<String, Any>,
-                    result = result)
+                method = call.method,
+                args = call.arguments as? Map<String, Any>,
+                result = result
+            )
         }
     }
 
     override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
         channel.setMethodCallHandler(null)
+        listenerHandler.dispose()
+        listenerHandler
     }
 
     override fun onAttachedToActivity(binding: ActivityPluginBinding) {
@@ -57,19 +64,20 @@ class ApphudPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
         setHeaders()
 
         handlers = listOf(
-                InitializationHandler(InitializationRoutes.stringValues(), context = this.context),
-                MakePurchaseHandler(MakePurchaseRoutes.stringValues(), activity = sActivity),
-                HandlePurchasesHandler(HandlePurchasesRoutes.stringValues(), context = this.context),
-                AttributionHandler(AttributionRoutes.stringValues()),
-                OtherHandler(OtherRoutes.stringValues(), context = this.context),
-                UserPropertiesHandler(UserPropertiesRoutes.stringValues(), context = this.context),
-                PaywallLogsHandler(PaywallLogsRoutes.stringValues(), context = this.context)
+            InitializationHandler(InitializationRoutes.stringValues(), context = this.context),
+            MakePurchaseHandler(MakePurchaseRoutes.stringValues(), activity = sActivity),
+            HandlePurchasesHandler(HandlePurchasesRoutes.stringValues(), context = this.context),
+            AttributionHandler(AttributionRoutes.stringValues()),
+            OtherHandler(OtherRoutes.stringValues(), context = this.context),
+            UserPropertiesHandler(UserPropertiesRoutes.stringValues(), context = this.context),
+            PaywallLogsHandler(PaywallLogsRoutes.stringValues(), context = this.context),
+            PromotionalsHandler(PromotionalsRoutes.stringValues())
         )
     }
 
     private fun setHeaders() {
-        HttpUrlConnectionExecutor.X_SDK = "flutter"
-        HttpUrlConnectionExecutor.X_SDK_VERSION = "2.1.0"
+        HeadersInterceptor.X_SDK = "Flutter"
+        HeadersInterceptor.X_SDK_VERSION = BuildConfig.APPHUD_FLUTTER_SDK_VERSION
     }
 
     override fun onDetachedFromActivityForConfigChanges() {