+++ example/lib/src/feature/home/sku_details_widget.dart
@@ -5,12 +5,14 @@ import 'package:flutter/material.dart';
 class SkuDetailsWidget extends StatelessWidget {
   final SkuDetailsWrapper? skuDetails;
   final VoidCallback? onTap;
+  final VoidCallback? onPromote;
   final bool wrapInCard;
 
   const SkuDetailsWidget({
     Key? key,
     this.skuDetails,
     this.onTap,
+    this.onPromote,
     bool? wrapInCard,
   })  : wrapInCard = wrapInCard ?? true,
         super(key: key);
@@ -21,13 +23,20 @@ class SkuDetailsWidget extends StatelessWidget {
       return _wrapInCard(child: ListTile(title: Text('skuDetails is null')));
     }
     final SkuDetailsWrapper skuDetailsLocal = skuDetails!;
-    return InkWellStack(
+    return InkWell(
       onTap: onTap,
       child: _wrapInCard(
         child: ListTile(
           title: Text('${skuDetailsLocal.title} (${skuDetailsLocal.sku})'),
           leading: Text(skuDetailsLocal.price),
           subtitle: Text(skuDetailsLocal.description),
+          trailing: HeroMode(
+            enabled: false,
+            child: FloatingActionButton(
+              onPressed: onPromote,
+              child: Text('P'),
+            ),
+          ),
         ),
       ),
     );