+++ example/lib/src/feature/initialization/initialization_bloc.dart
@@ -1,8 +1,11 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:apphud/apphud.dart';
+import 'package:apphud/models/apphud_models/apphud_debug_level.dart';
+import 'package:apphud/models/apphud_models/apphud_non_renewing_purchase.dart';
 import 'package:apphud/models/apphud_models/apphud_paywalls.dart';
+import 'package:apphud/models/apphud_models/apphud_subscription.dart';
+import 'package:apphud/models/apphud_models/composite/apphud_product_composite.dart';
 import 'package:apphud_example/src/feature/common/app_secrets_base.dart';
 import 'package:apphud_example/src/feature/common/debug_print_mixin.dart';
 import 'package:apphud_example/src/feature/navigation/navigation_bloc.dart';
@@ -13,7 +16,8 @@ import 'initialization_event.dart';
 import 'initialization_state.dart';
 
 class InitializationBloc extends Bloc<InitializationEvent, InitializationState>
-    with DebugPrintMixin {
+    with DebugPrintMixin
+    implements ApphudListener {
   final AppSecretsBase _appSecrets;
   final NavigationBloc _navigationBloc;
 
@@ -25,14 +29,7 @@ class InitializationBloc extends Bloc<InitializationEvent, InitializationState>
         super(InitializationState.trying());
 
   void _fetchPaywalls() async {
-    try {
-      final ApphudPaywalls paywalls = await Apphud.getPaywalls();
-      add(InitializationEvent.paywallsFetchSuccess(paywalls));
-      printAsJson('getPaywalls()', paywalls);
-    } catch (e) {
-      add(InitializationEvent.paywallsFetchFailure(e.toString()));
-      printError('getPaywalls()', e);
-    }
+    unawaited(Apphud.setListener(listener: this));
   }
 
   @override
@@ -41,14 +38,13 @@ class InitializationBloc extends Bloc<InitializationEvent, InitializationState>
   ) =>
       event.map(
         initializeTrying: _mapInitializeTrying,
-        paywallsFetchFailure: _mapPaywallsFetchFailure,
         paywallsFetchSuccess: _mapPaywallsFetchSuccess,
       );
 
   Stream<InitializationState> _mapInitializeTrying(
       InitializeTrying value) async* {
     try {
-      await Apphud.enableDebugLogs();
+      await Apphud.enableDebugLogs(level: ApphudDebugLevel.high);
       await Apphud.startManually(
         apiKey: _appSecrets.apiKey,
         userID: _appSecrets.userID,
@@ -63,17 +59,10 @@ class InitializationBloc extends Bloc<InitializationEvent, InitializationState>
         },
       );
 
-      _paywallsIos();
-      _paywallsDidLoadCallbackIos();
       _fetchPaywalls();
     } catch (e) {
       yield InitializationState.startFail(e.toString());
     }
-  }
-
-  Stream<InitializationState> _mapPaywallsFetchFailure(
-      PaywallsFetchFailure event) async* {
-    yield InitializationState.paywallsFetchFail(event.error);
   }
 
   Stream<InitializationState> _mapPaywallsFetchSuccess(
@@ -96,27 +85,40 @@ class InitializationBloc extends Bloc<InitializationEvent, InitializationState>
     );
   }
 
-  void _paywallsIos() {
-    if (Platform.isIOS) {
-      Apphud.paywalls().then(
-        (value) => printAsJson(
-          'paywalls',
-          value,
-        ),
-        onError: (e) => printError('paywalls', e),
-      );
-    }
+  @override
+  Future<void> apphudDidChangeUserID(String userId) async {
+    printAsJson('ApphudListener.apphudDidChangeUserID', userId);
   }
 
-  void _paywallsDidLoadCallbackIos() {
-    if (Platform.isIOS) {
-      Apphud.paywallsDidLoadCallback().then(
-        (value) => printAsJson(
-          'paywallsDidLoadCallback()',
-          value,
-        ),
-        onError: (e) => printError('paywallsDidLoadCallback()', e),
-      );
-    }
+  @override
+  Future<void> apphudDidFecthProducts(
+    List<ApphudProductComposite> products,
+  ) async {
+    printAsJson('ApphudListener.apphudDidFecthProducts', products);
+  }
+
+  @override
+  Future<void> paywallsDidFullyLoad(ApphudPaywalls paywalls) async {
+    printAsJson('ApphudListener.paywallsDidFullyLoad', paywalls);
+    add(InitializationEvent.paywallsFetchSuccess(paywalls));
+  }
+
+  @override
+  Future<void> paywallsDidLoad(ApphudPaywalls paywalls) async {
+    printAsJson('ApphudListener.paywallsDidLoad', paywalls);
+  }
+
+  @override
+  Future<void> apphudNonRenewingPurchasesUpdated(
+    List<ApphudNonRenewingPurchase> purchases,
+  ) async {
+    printAsJson('ApphudListener.apphudNonRenewingPurchasesUpdated', purchases);
+  }
+
+  @override
+  Future<void> apphudSubscriptionsUpdated(
+    List<ApphudSubscriptionWrapper> subscriptions,
+  ) async {
+    printAsJson('ApphudListener.apphudSubscriptionsUpdated', subscriptions);
   }
 }